% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocr.gen.R
\name{add_toc}
\alias{add_toc}
\title{Add a table of contents (TOC) to an (R) Markdown document}
\usage{
add_toc(
  md,
  min_tier = 2L,
  max_tier = 6L,
  position = "above",
  md_flavor = c("github", "gitlab"),
  add_title = TRUE,
  title = "Table of contents",
  title_tier = min_tier,
  add_backlinks = add_title,
  backlink_strings = c("↑", "↓"),
  backlink_position = c("before", "after"),
  listing_style = c("-", "*", "ordered", "indented"),
  toc_id = "toc",
  old_toc_id = toc_id
)
}
\arguments{
\item{md}{(R) Markdown document to be processed as a single file path, a single URL or a character vector (one string per line).}

\item{min_tier}{Minimum tier of headers (\verb{<h1>} – \verb{<h6>}) to include in the TOC. Integer between \code{1} and \code{6}. \code{min_tier = 2} for example means to create TOC
entries for all \verb{<h2>} and below headers.}

\item{max_tier}{Maximum tier of headers (\verb{<h1>} – \verb{<h6>}) to include in the TOC. Integer between \code{1} and \code{6}. \code{max_tier = 5} for example means to create TOC
entries for all headers down to \verb{<h5>}. \code{max_tier} must be >= \code{min_tier}.}

\item{position}{Position in the Markdown document at which to add the TOC. Possible values include:
\itemize{
\item \code{"top"}: The very beginning of the document, i.e. the first line.
\item \code{"bottom"}: The very end of the document, i.e. the last line.
\item \code{"above"}: Above the lines between the uppermost header of tier <= \code{min_tier} and the next header above (if any).
\item \code{"below"}: Below the lines between the uppermost header of tier <= \code{min_tier} and the next header above (if any), i.e. right above the uppermost header
of tier <= \code{min_tier}.
\item \code{"none"}: Only remove a possibly existing TOC.
\item A line number, given as a positive integer.
}}

\item{md_flavor}{Markdown flavor. Possible values include:
\itemize{
\item \code{"github"}:
\item \code{"gitlab"}:
}}

\item{add_title}{Include a TOC title? Logical. Note that no backlinks are
added at all if \code{add_title = FALSE} and no header line are found above
\code{position}, regardless of \code{add_backlinks = TRUE}.}

\item{title}{Title of the TOC. A character scalar.}

\item{title_tier}{Tier/formatting of TOC title. Possible values include:
\itemize{
\item An integer between \code{1L} and \code{6L} representing the \verb{<h1>} – \verb{<h6>} tier.
\item \code{"regular"}: Simple unformatted non-header text.
\item \code{"bold"}: Bold (\verb{<strong>}) non-header text.
\item \code{"italic"}: Italic (\verb{<em>}) non-header text.
}}

\item{add_backlinks}{Add a link back to the TOC to each Markdown header. A logical scalar. Note that if \code{add_backlinks = TRUE} and \code{add_title = FALSE}, as a
fallback the backlinks point to the next header line above \code{position} (if any found). This will also be the case if \code{md_flavor = "gitlab"} and
\code{title_tier} is set to a non-header value (\code{"regular"}, \code{"bold"} or \code{"italic"}) because GitLab currently ignores manually set HTML \verb{<id>} attributes.}

\item{backlink_strings}{String(s) to use as link text back to the TOC. A character vector of length 1 or 2. If two strings are provided, the first one will
be used for backlinks below \code{position}, the second one for backlinks above \code{position}. Note that at least Unicode 7.0 support is required for the default
symbols \verb{\\U1F805} and \verb{\\U1F807} to be correctly displayed.}

\item{backlink_position}{Position of the backlinks. Possible values include:
\itemize{
\item \code{"before"}: Before the actual header text.
\item \code{"after"}: After the actual header text.
}}

\item{listing_style}{Format to use for listing the TOC entries. Possible values include:
\itemize{
\item \code{"-"}: Create an unordered list using a hyphen as listing symbol.
\item \code{"*"}: Create an unordered list using an asterisk as listing symbol.
\item \code{"ordered"}: Create an ordered list using \code{1}, \code{2}, \code{3}, ...  as listing symbols.
\item \code{"indented"}: Use non-breaking spaces (\verb{&nbsp;}) to create visual indentation. Useful if the headers are already numbered.
}}

\item{toc_id}{HTML \verb{<id>} attribute of the TOC title if \code{title_tier} is set to a non-header value (\code{"regular"}, \code{"bold"} or \code{"italic"}). A character scalar.}

\item{old_toc_id}{HTML \verb{<id>} attribute of the old TOC title (in order to have old backlinks with an ID other than \code{toc_id} removed). A character scalar.}
}
\value{
The processed Markdown document as a character vector (one string per line).
}
\description{
DESCRIPTION Blabla...
}
\details{
This function tries to adhere to the \href{http://commonmark.org/}{CommonMark specification}, i.e. to interpret the Markdown syntax the same way as the
\href{https://github.com/jgm/commonmark.js}{commonmark.js} reference implementation under <try.commonmark.org> does.
}
\examples{
md <- paste0("https://raw.githubusercontent.com/ropensci/pdftools/",
             "e7248d9956c7e73968628fa3a8ed37f0a8c23b37/README.md")

md |>
  tocr::add_toc(position = 9) |>
  pal::cat_lines()

md |>
  tocr::add_toc(listing_style = "ordered") |>
  pal::cat_lines()
}
