% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocr.gen.R
\name{convert_header_to_anchor}
\alias{convert_header_to_anchor}
\title{Convert header text to HTML anchor}
\usage{
convert_header_to_anchor(
  header_text,
  md_flavor,
  preceding_anchor_links = character(),
  backlink_string = character(),
  backlink_position
)
}
\description{
Convert header text to HTML anchor
}
\keyword{internal}
