% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocr-GEN.R
\name{evaluate_position}
\alias{evaluate_position}
\title{Evaluate \code{position}}
\usage{
evaluate_position(position, position_offset, md_lines, min_tier, filename)
}
\description{
Evaluate \code{position}
}
\keyword{internal}
