% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocr-GEN.R
\name{get_non_code_block_indices}
\alias{get_non_code_block_indices}
\title{Get non-code-block indices}
\usage{
get_non_code_block_indices(md_lines)
}
\description{
Get non-code-block indices
}
\keyword{internal}
