% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ua.R
\name{get_ua}
\alias{get_ua}
\title{get_ua}
\usage{

  get_ua(url = "https://developers.whatismybrowser.com/useragents/explore/software_name/",
  browser = c("chrome", "firefox"))
}
\arguments{
\item{url}{Website mit Liste an user agents}

\item{browser}{Browser für die user agents abgefragt werden sollen}
}
\value{
Liste mit user agents
}
\description{
ruft eine Liste mit user agents ab
}
