% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocr.gen.R
\name{handle_existing_toc}
\alias{handle_existing_toc}
\title{Handle existing TOC (integrity checks and removal)}
\usage{
handle_existing_toc(md_lines, non_code_block_indices, position, filename)
}
\description{
Handle existing TOC (integrity checks and removal)
}
\keyword{internal}
