% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_crawl.R
\name{hidden_crawl}
\alias{hidden_crawl}
\title{hidden crawler}
\usage{
hidden_crawl(url, ua = "", proxy = "socks5h://127.0.0.1:9153", hidden = T,
  info = F, verbose = F, timeout = 5, encoding = "UTF-8", wait = 0,
  try.con = 10, ssl = F)
}
\arguments{
\item{url}{zu crawlende url}

\item{ua}{zu verwendender user agent}

\item{proxy}{hidden=T, verwendeter socks port, um TOR-Browser anzusprechen}

\item{hidden}{crawl mit eigener (hidden=F) oder TOR-IP (hidden=T)}

\item{info}{anzeige welche IP verwendet wird (T/F)}

\item{verbose}{Ausgabe zus채tzlicher Informationen zum Crawl (T/F)}

\item{timeout}{Zeit, in der auf R체ckmeldung von Website gewartet wird; vor Verbindungsabbruch}

\item{encoding}{verwendeter Zeichensatz der Website}

\item{wait}{Wartezeit vor Crawl in Sekunden}

\item{try.con}{Anzahl der gescheiterten Verbindungsversuche mit TOR, bis eigene IP verwendet wird}

\item{ssl}{ssl.verifypeer (T/F)}
}
\value{
XML-Inhalt der aufgerufenen url
}
\description{
crawlt den Inhalt einer Website und gibt den passenden xml-Output zur체ck
Der Crawl 체ber die eigene IP oder das TOR-Netzwerk gestartet werden.
}
