% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocr.gen.R
\name{process_md}
\alias{process_md}
\title{Process Markdown document line by line (clean it and build new TOC entries)}
\usage{
process_md(
  md_lines,
  non_code_block_indices,
  min_tier,
  max_tier,
  md_flavor,
  add_backlinks,
  backlink_strings,
  backlink_position,
  toc_id,
  old_toc_id,
  new_toc_id,
  new_toc_id_position,
  use_fallback,
  is_header_title
)
}
\description{
Process Markdown document line by line (clean it and build new TOC entries)
}
\keyword{internal}
