% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocr.gen.R
\name{remove_toc}
\alias{remove_toc}
\title{Remove an auto-generated table of contents (TOC) from an (R) Markdown document}
\usage{
remove_toc(md, old_toc_id = "toc")
}
\arguments{
\item{md}{(R) Markdown document to be processed as a single file path, a single URL or a character vector (one string per line).}

\item{old_toc_id}{HTML \verb{<id>} attribute of the old TOC title (in order to have old backlinks with an ID other than \code{toc_id} removed). A character scalar.}
}
\value{
The processed Markdown document as a character vector (one string per line).
}
\description{
This is a simple convenience wrapper for \code{add_toc(..., position = "none")}.
}
\examples{
\dontrun{
tocr::remove_toc(md = paste0("https://raw.githubusercontent.com/thlorenz/doctoc/",
                             "1d386261972d35c6bcd187d0a00e666f9d893d8d/README.md")) |>
  cat(sep = "\n")}
}
\seealso{
\code{\link[=add_toc]{add_toc()}}
}
