% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocr-package.R
\docType{package}
\name{tocr-package}
\alias{tocr}
\alias{tocr-package}
\title{tocr: TOC Generation for (R) Markdown Documents}
\description{
tocr allows to automatically add/update/remove a table of contents (TOC) to/from an (R) Markdown document. Internal links are compatible with HTML "anchors", commonly referred to as "# links" pointing to HTML fragment identifiers generated by sites like GitHub or GitLab when rendering the Markdown. tocr supports backlinks to ease navigation within large documents as well as other options to tailor the TOC to specific needs. It can also convert a wide range of similar auto-generated TOC formats to its own one.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/salim_b/r/pkgs/tocr}
  \item Report bugs at \url{https://gitlab.com/salim_b/r/pkgs/tocr/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
